third_party/blink/web_tests/platform/mac-mac11-arm64/external/wpt/html/browsers/the-window-object/named-access-on-the-window-object/window-named-properties-expected.txt

This is a testharness.js-based test.
PASS Static name
PASS Static id
PASS Static name on the prototype
FAIL constructor assert_false: gsp.hasOwnProperty("constructor") expected false got true
FAIL duplicate property names assert_equals: expected 1 but got 0
PASS Dynamic name
PASS Ghost name
Harness: the test ran to completion.
